users gui for admin panel added

+++ controllers/handlerFactory.js
@@ -1,10 +1,16 @@
 const catchAsync = require('../utils/catchAsync');
 const AppError = require('../utils/appError');
 const APIFeatures = require('../utils/apiFeatures');
+const User = require('../models/userModel');
 
 exports.deleteOne = (Model) =>
   catchAsync(async (req, res, next) => {
-    const doc = await Model.findByIdAndDelete(req.params.id);
+    let doc;
+    if (Model === User) {
+      doc = await Model.findByIdAndUpdate(req.params.id, { active: false });
+    } else {
+      doc = await Model.findByIdAndDelete(req.params.id);
+    }
 
     if (!doc) {
       return next(new AppError('No document found with that ID', 404));
@@ -15,6 +21,7 @@ exports.deleteOne = (Model) =>
 
 exports.updateOne = (Model) =>
   catchAsync(async (req, res, next) => {
+    if (req.file) req.body.photo = req.file.filename;
     const doc = await Model.findByIdAndUpdate(req.params.id, req.body, {
       new: true,
       runValidators: true,
@@ -29,6 +36,8 @@ exports.updateOne = (Model) =>
 
 exports.createOne = (Model) =>
   catchAsync(async (req, res, next) => {
+    if (req.file) req.body.photo = req.file.filename;
+    else req.body.photo = 'default.jpg';
     const newDoc = await Model.create(req.body);
 
     res.status(201).json({
